添加用户id

+++ purchasing/src/main/java/com/team/purchasing/bean/Proclamation.java
@@ -37,13 +37,13 @@ public class Proclamation implements Serializable {
     private String createTime;
 
     @ApiModelProperty(value = "创建人")
-    private Integer createUserId;
+    private Long createUserId;
 
     @ApiModelProperty(value = "更新时间")
     private String updateTime;
 
     @ApiModelProperty(value = "更新人")
-    private Integer updateUserId;
+    private Long updateUserId;
 
     @ApiModelProperty(value = "分页")
     private Page page;
