+++ purchasing/src/main/java/com/team/purchasing/bean/ProductCategory.java
@@ -34,12 +34,12 @@ public class ProductCategory implements Serializable {
     private String createTime;
 
     @ApiModelProperty(value = "创建用户")
-    private Integer createUserId;
+    private Long createUserId;
 
     @ApiModelProperty(value = "更新时间")
     private String updateTime;
 
     @ApiModelProperty(value = "更新用户")
-    private Integer updateUserId;
+    private Long updateUserId;
 
 }
