+++ purchasing/src/main/java/com/team/purchasing/bean/booking/OrderBooking.java
@@ -25,11 +25,11 @@ public class OrderBooking {
 
     /** 用户id **/
     @ApiModelProperty(value = "用户id")
-    private Integer userId;
+    private Long userId;
 
     /** 医院或者院校id **/
     @ApiModelProperty(value = "医院或者院校id")
-    private Integer HCId;
+    private Long HCId;
 
     /** 订单编号 **/
     @ApiModelProperty(value = "订单编号")
@@ -81,7 +81,7 @@ public class OrderBooking {
 
     /** 创建用户id **/
     @ApiModelProperty(value = "创建用户id")
-    private Integer createUserId;
+    private Long createUserId;
 
     /** 最近一次修改时间 **/
     @ApiModelProperty(value = "最近一次修改时间")
@@ -89,7 +89,7 @@ public class OrderBooking {
 
     /** 最近修改用户id **/
     @ApiModelProperty(value = "最近修改用户id")
-    private Integer updateUserId;
+    private Long updateUserId;
 
     @ApiModelProperty(value = "下单来源")
     private Integer orderFrom;
